feat: C2PA embed/trust, VC export FFI, velocity analysis, forensic HTML report, store integrity checks
- authorproof-protocol: C2PA embed.rs and trust.rs modules, builder enhancements, new types
- cpoe FFI: vc_export module, evidence_derivative expansion, ephemeral/sentinel/types updates
- fingerprint: velocity_analysis, global profile, digraph analysis, 243 new test lines
- forensics: advanced_metrics fix
- platform: keystroke event additions across macOS/Linux/Windows/synthetic
- sentinel: core_session and ipc_handler improvements
- store: integrity checks, new store operations
- war/profiles: VC profile enhancements
- report: forensic HTML report template

+++ crates/cpoe/src/forensics/advanced_metrics.rs
@@ -302,7 +302,7 @@ pub fn compute_fatigue_phase_binding(
     let q_warmup = (metrics.warmup_fraction * 20.0).round() as u8;
     let q_plateau = (metrics.plateau_fraction * 20.0).round() as u8;
     let q_fatigue = (metrics.fatigue_fraction * 20.0).round() as u8;
-    h.update([q_warmup, q_plateau, q_fatigue, metrics.dominant_phase as u8]);
+    h.update([q_warmup, q_plateau, q_fatigue, metrics.dominant_phase]);
     h.finalize().into()
 }
